+++ src/main/java/com/umc/TheGoods/domain/order/Cart.java
@@ -2,7 +2,6 @@ package com.umc.TheGoods.domain.order;
 
 import com.umc.TheGoods.domain.common.BaseDateTimeEntity;
 import com.umc.TheGoods.domain.item.Item;
-import com.umc.TheGoods.domain.item.ItemOption;
 import com.umc.TheGoods.domain.member.Member;
 import lombok.*;
 
@@ -20,9 +19,6 @@ public class Cart extends BaseDateTimeEntity {
     @Column(name = "cart_id")
     private Long id;
 
-    @Column(nullable = false)
-    private Integer amount;
-
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", nullable = false)
     private Member member;
@@ -30,10 +26,6 @@ public class Cart extends BaseDateTimeEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "item_id", nullable = false)
     private Item item;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "item_option_id")
-    private ItemOption itemOption;
 
     // 연관관계 메소드
     public void setMember(Member member) {
@@ -52,16 +44,5 @@ public class Cart extends BaseDateTimeEntity {
         item.getItemCartList().add(this);
     }
 
-    public void setItemOption(ItemOption itemOption) {
-        if (this.itemOption != null) {
-            this.itemOption.getCartList().remove(this);
-        }
-        this.itemOption = itemOption;
-        itemOption.getCartList().add(this);
-    }
 
-    public Cart updateAmount(Integer amount) {
-        this.amount = amount;
-        return this;
-    }
 }
